// src/generated/models/Ai_alhasbakpihistoriesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Ai_alhasbakpihistoriesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Ai_alhasbakpihistoriesstatecode = keyof typeof Ai_alhasbakpihistoriesstatecode;
export const Ai_alhasbakpihistoriesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Ai_alhasbakpihistoriesstatuscode = keyof typeof Ai_alhasbakpihistoriesstatuscode;

export interface Ai_alhasbakpihistoriesBase {
  ai_actual?: number;
  ai_alhasbakpihistoryid: string;
  ai_id?: number;
  ai_kpi_id?: string;
  ai_name: string;
  ai_period?: string;
  ai_target?: number;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Ai_alhasbakpihistoriesstatecode;
  statuscode?: Ai_alhasbakpihistoriesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Ai_alhasbakpihistories extends Ai_alhasbakpihistoriesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}